+++ print.h
@@ -8,7 +8,7 @@
 #include "stm32f4xx_ll_gpio.h"
 #include <stddef.h> 
 
-void UART_config (size_t peripheral_clock, size_t UART_baudrate);
+void UART_config (uint32_t peripheral_clock, uint32_t UART_baudrate);
 void print (const char* ptr); // UART debug print
 
 #endif /* PRINT_H */
